requirements.txt
# Core libraries
streamlit==1.35.0
openai==1.78.1
PyPDF2==3.0.1
python-dotenv==1.0.1
pdfplumber==0.11.6

azure-core==1.33.0
azure-identity==1.21.0
azure-common==1.1.28

azure-search-documents==11.5.2

uvicorn==0.29.0

anyio==4.9.0
annotated-types==0.7.0
attrs==25.3.0
nest_asyncio
azure-storage-blob
beautifulsoup4
rapidfuzz>=3.0.0
python-docx
reportlab